+++ Week7/Lecture19/go.mod
@@ -1,0 +1,3 @@
+module reverseString
+
+go 1.18
