+++ Week7/Lecture19/main.go
@@ -3,28 +3,11 @@ package main
 import (
 	"io"
 	"os"
-	"strings"
+	"reverseString/reverseString"
 )
 
-type ReverseStringReader io.Reader
-
 func main() {
-	io.Copy(os.Stdout, *NewReverseStringReader("Hello world"))
-}
-
-func NewReverseStringReader(input string) *ReverseStringReader {
-	var reader ReverseStringReader
-
-	runes := []rune(input)
-	for i, j := 0, len(runes)-1; i < j; i, j = i+1, j-1 {
-		runes[i], runes[j] = runes[j], runes[i]
-	}
-	turnToStrings := string(runes)
-
-	toReader := strings.NewReader(turnToStrings)
-
-	reader = toReader
-	return &reader
+	io.Copy(os.Stdout, *reverseString.NewReverseStringReader("Hello world"))
 }
 
 // Output
